requirements.txt
asgiref==3.5.2
backports.zoneinfo;python_version<"3.9"
certifi==2022.6.15.1
charset-normalizer==2.1.1
Django==4.1.1
django-cors-headers==3.13.0
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
Pillow==9.2.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
